requirements.txt
# Second Brain - Python Dependencies
# Install: pip install -r requirements.txt

# Core server framework
fastapi>=0.128.0
uvicorn>=0.40.0
pydantic>=2.12.5

# Claude SDK
claude-agent-sdk>=0.1.36
mcp>=1.25.0

# LLM providers
anthropic>=0.78.0
google-genai>=1.62.0
perplexityai>=0.26.0

# Google APIs
google-api-python-client>=2.188.0
google-auth>=2.47.0
google-auth-oauthlib>=1.2.4

# HTTP clients
requests>=2.32.5
urllib3>=2.6.3
httpx>=0.28.1

# Data / ML
numpy>=2.4.1
sentence-transformers>=5.2.0
faiss-cpu>=1.13.2

# Web scraping / parsing
beautifulsoup4>=4.14.3
readability-lxml>=0.8.4.1
markdownify>=1.2.2
pypdf>=6.6.0

# Finance
plaid-python>=38.0.0

# Chess engine
chess>=1.11.2

# Utilities
PyYAML>=6.0.3
python-dotenv>=1.2.1
filelock>=3.20.3

# Optional (uncomment if needed)
# flask>=3.0              # Plaid OAuth callback server
# playwright>=1.0         # Headless browser for doc scraping
